added scale_umap_data function, umap is optional
